requirements.txt
Flask
kafka-python
opentelemetry-api==0.6b0
opentelemetry-sdk==0.6b0
opentelemetry-ext-flask==0.6b0
opentelemetry-ext-jaeger==0.6b0
opentelemetry-ext-lightstep==0.2a0

